Moved code to other src files
