--- src/constants.js
@@ -1,1 +1,0 @@
-export var SCOPE_BOUNDARY = "__d3TemplateScope";
